delete old package and fix include old classes
