+++ .settings/org.eclipse.buildship.core.prefs
@@ -1,6 +1,5 @@
 build.commands=org.eclipse.jdt.core.javabuilder
 connection.arguments=
-connection.gradle.distribution=GRADLE_DISTRIBUTION(WRAPPER)
 connection.gradle.user.home=null
 connection.java.home=null
 connection.jvm.arguments=
